Sweep 0.1.102: Haunt completion pass, personal loot v2, LFG role shape, contagion parity, attunement suffix stats
Warlock:
- Haunt auto-apply now teaches+uses Unstable Affliction (talent chain
  30108-47843 granted via CLASS_PERK_GRANTS so spec-switch revokes it)
- Curse of Agony and Curse of the Elements coexist: Aura::CanStackWith
  rejected any same-caster SPELL_SPECIFIC_CURSE; narrow exception for the
  980<->1490 pair (core-patch 0034)
- Haunt castable while moving: server side was already instant; the
  client DBC blocked the cast - build_patch.py now sets CastingTimeIndex=1
  and clears the movement-interrupt bit on Haunt ranks (ships with client)
- Affliction DoTs 4x -> 21x (2000%) per report 122

Druid Balance:
- Starfall hits: 30% chance of a free Moonfire on the target, which
  cascades a standalone Hurricane (constant hoisted above its use site)
- Insect Swarm is now contagious: carrier-based spread + refresh on
  re-apply, all ranks recognized (report 119)

Mage Fire:
- Living Bomb spreads carrier-to-carrier like the warlock contagion
  (60y search, 15y hops) instead of only radiating from the mage (128)

Loot:
- Personal loot v2: drop table rolls once, every item flagged freeforall
  so each member loots the full table independently - no overlap, no
  rolls, gold splits once, skinning stays single-taker (126). Latecomer
  lazy view fill + unlootedCount widened to uint32. Core-patch 0029 V2.

LFG:
- Untalented bots defaulted to DPS tabs and could not be respec'd into
  tank/healer, and the fill counted talent-guess roles instead of
  declared ones -> 5-dps groups (129). Untalented defaults are now
  prot/prot/feral/resto and the fill reads declared LFG roles
  (core-patch 0035).

Rogue:
- ScanLockpick rogue-only gate removed: account-wide lockpicking now
  auto-unlocks chests and bagged lockboxes on any character (135)

Party XP:
- Bots were counted in the party max bonus: RNDBOT1 carries x2.5 and
  raised every grouped human to match (136). Max now counts humans only.

Attunement:
- Suffix-only items (e.g. Tumultuous Necklace) never attuned: both
  writers read template stats only. Suffix stats now folded into
  attunement + curator banking ratchet (127)

Loot bind warnings (121): not a server bug - client itemcache.wdb is
stale; wiping the client Cache folder clears it

+++ src/server/game/Loot/LootMgr.h
@@ -321,7 +321,7 @@ struct Loot
     std::vector<LootItem> items;
     std::vector<LootItem> quest_items;
     uint32 gold;
-    uint8 unlootedCount{0};
+    uint32 unlootedCount{0};
     ObjectGuid roundRobinPlayer;        // GUID of the player having the Round-Robin ownership for the loot. If 0, round robin owner has released.
     ObjectGuid lootOwnerGUID;
     LootType loot_type{LOOT_NONE};      // required for achievement system
